feat(demo/fluid-cursor-advanced): add interactive container picker
